% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adjustedL}
\alias{adjustedL}
\title{Adjusted Transect Length}
\usage{
adjustedL(transect, trunc)
}
\arguments{
\item{transect}{GIS file of the transect of class SpatialLines}

\item{trunc}{Truncation distance (meters)}
}
\value{
Length of the transect (meters) used by distance sampling packages to correctly calculate the covered area.
}
\description{
Popular distance sampling packages automatically calculate the area covered in the survey based on a straight line (L x 2w). This function
calculates the the area covered in a curved transects and provides the adjusted length necessary for input into these packages (curved covered area รท (2 ร w))
}
\examples{
adjustedL(transect=transect, trunc=100)
}
