% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cca}
\alias{cca}
\title{Adjusted Transect Length for Curved Covered Area Calculation}
\usage{
cca(transect, trunc)
}
\arguments{
\item{transect}{GIS file of the transect of class SpatialLines}

\item{trunc}{Truncation distance (meters)}
}
\value{
Length of the transect (meters) used by distance sampling packages to correctly calculate the covered area.
}
\description{
Adjusted Transect Length for Curved Covered Area Calculation
}
\examples{
cca(transect=transect, trunc=100)
}
