% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveTransect-package.R
\docType{package}
\name{CurveTransect}
\alias{CurveTransect}
\title{Curving Transect Distance Sampling}
\description{
\code{CurveTransect} facilitates distance sampling on a curving transect.
}
\details{
\code{observerXY}: creates points on a transect at specific meter intervals

\code{objectXY}: provides spatial coordinates for each detected object in a distance sampling survey

\code{nearestX}: measures the shortest distance between each detected object anda curving transect

\code{adjustedL}: calculates the an adjusted transect length for curved covered area calculations

Examples of how to use this package are available at
\url{https://github.com/esbach/CurveTransect}
}
\references{
Esbach M.S. and R.K. Patra. In review. Distance Sampling from Curving Transects in Dense Forests.
Methods in Ecology and Evolution.
}
\author{
Michael Esbach \href{mailto:mesbach@princeton.edu}{mesbach@princeton.edu}
}
