% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nearest.distance}
\alias{nearest.distance}
\title{Nearest Distance between Detected Object and Curving Transect}
\usage{
nearest.distance(detections, transect)
}
\arguments{
\item{detections}{Output of "objectXY": data-frame with the animal's spatial location as xy coordinates, which must be labelled: "x.obs", "y.obs"}

\item{transect}{GIS file of the transect of class SpatialLines}
}
\value{
Data-frame with measures of distance (meters) between each detection and the nearest location on the transect
}
\description{
Nearest Distance between Detected Object and Curving Transect
}
\examples{
distance = nearest.distance(detections = detections, transect = transect)
}
