% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nearestX}
\alias{nearestX}
\title{Nearest Distance between Detected Object and Curving Transect}
\usage{
nearestX(detections, transect)
}
\arguments{
\item{detections}{Output of "objectXY": data-frame with the animal's spatial location as xy coordinates, which must be labelled: "x.obs", "y.obs"}

\item{transect}{GIS file of the transect of class SpatialLines}
}
\value{
Data-frame with measures of distance (meters) between each detection and the nearest location on the transect
}
\description{
This function uses the output of “objectXY” (xy coordinates of each detected object) and a GIS file of the transect to measure the nearest distance between each object and the curved transect. It returns a data-frame of distances (x) that can be used in distance sampling analyses.
}
\examples{
nearest.distance(detections = detections, transect = transect)
}
