% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveTransect.R
\name{objectXY}
\alias{objectXY}
\title{Location of Detected Objects}
\usage{
objectXY(transectXY, detections, buffer)
}
\arguments{
\item{transectXY}{Two column data-matrix containing the transect's spatial coordinates}

\item{detections}{Data frame with colums 'meter,' 'distance,' and 'angle.'}

\item{buffer}{Number of meters around the observer's location used to make bearing on the transect.}
}
\value{
Data-frame with the animal's spatial location as xy coordinates, plus the original 'meter', 'distance,' and 'angle.'
}
\description{
Location of Detected Objects
}
\examples{
objectXY(transect=transectXY, detections=ds, buffer=5) 
}
