% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveTransect.R
\name{observerXY}
\alias{observerXY}
\title{Equidistant Points on Transect}
\usage{
observerXY(transectXY, spacing)
}
\arguments{
\item{transectXY}{Two column data-matrix containing the transect's spatial coordinates}

\item{spacing}{Numeric distance (in meters) between each point on the transect}
}
\value{
Data-frame with spatial coordinates for each specified point on the transect
}
\description{
Equidistant Points on Transect
}
