% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveTransect-package.R
\docType{package}
\name{CurveTransect}
\alias{CurveTransect}
\title{Curving Transect Distance Sampling}
\description{
`CurveTransect` facilitates distance sampling on a curving transect.
}
\details{
Examples of how to use this package are available at
https://github.com/esbach/CurveTransect.

For help with distance sampling, there is a Google Group
https://groups.google.com/forum/#!forum/distance-sampling.
}
\references{
Esbach M.S. and R.K. Patra. In review. Distance Sampling from Curving Transects in Dense Forests.
  Methods in Ecology and Evolution.
}
\author{
Michael Esbach <mesbach@princeton.edu>
}
