% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{observerXY}
\alias{observerXY}
\title{Equidistant Points on Transect}
\usage{
observerXY(transect, spacing)
}
\arguments{
\item{spacing}{Numeric spacing (in meters) between each point on the transect}

\item{transectXY}{Two column data-matrix containing the transect's spatial coordinates}
}
\value{
Data-frame with spatial coordinates for each specified point on the transect
}
\description{
This function takes a GIS file of a transect and places a point at every meter (or every x meters depending not the spacing you choose). It returns a data-frame with xy coordinates for each meter.
}
\examples{
observerXY(transect=transectXY, spacing=1)
}
